# docs/api_specs.yaml
openapi: 3.0.3
info:
    title: Echo API
    version: 1.0.0
    description: |
        Local-first REST API for Echo - a multimodal semantic second brain.
        All processing runs on the user's device. No external network calls are
        made during normal operation. The API never generates content; it only
        stores, retrieves and surfaces the user's own existing notes.
servers:
    - url: http://127.0.0.1:8000/
      description: Local development server

#  REUSABLE COMPONENTS
components:
    schemas:
...
        Tag:
            type: object
            required: [id, name, created_at]
            properties:
                id:
                    type: string
                    format: uuid
                name:
                    type: string
                parent_id:
                    type: string
                    format: uuid
                    nullable: true
                    description: >
                        Null for top-level tags. If a parent tag is deleted,
                        this becomes null (orphaned to top level) rather than
                        the child being deleted.
                created_at:
                    type: string
                    format: date-time
...
        Error400:
            type: object
            required: [error, detail]
            properties:
                error:
                    type: string
                    example: "validation_error"
                detail:
                    type: string
                    example: "note_type must be one of: text, image, audio, canvas"
        Error404:
            type: object
            required: [error, detail]
            properties:
                error:
                    type: string
                    example: "not_found"
                detail:
                    type: string
                    example: "Note with id 'a1b2c3d4...' does not exist."
...
        Error500:
            type: object
            required: [error, detail]
            properties:
                error:
                    type: string
                    example: "internal_error"
                detail:
                    type: string
                    example: "An unexpected error occurred."

    #  PARAMETERS
    parameters:
        NoteId:
            name: id
            in: path
            required: true
            schema:
                type: string
                format: uuid
            description: The UUID of the note.
...
paths:
...
    /tags:
        post:
            operationId: createTag
            summary: Create a manual tag
            description: >
                Creates a new tag.
            tags: [Tags]
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required: [name]
                            properties:
                                name:
                                    type: string
                                parent_id:
                                    type: string
                                    format: uuid
                                    nullable: true
            responses:
                "201":
                    description: Tag created.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Tag"
                "400":
                    description: >
                        Validation error, e.g. Parent_id references a
                        non-existent tag.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error400"
                "500":
                    description: Internal server error.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error500"

        get:
            operationId: listTags
            summary: List all tags
            description: Returns all tags (flat list).
            tags: [Tags]
            responses:
                "200":
                    description: Array of tags.
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/Tag"
                "500":
                    description: Internal server error.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error500"
    /tags/{id}:
        patch:
            operationId: updateTag
            summary: Rename a tag or change its parent
            description: >
                Used both for renaming and for moving a tag to a different
                point in the hierarchy (set parent_id to null to make it
                top-level). Re-parenting is independent of deletion.
            tags: [Tags]
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: string
                      format: uuid
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                name:
                                    type: string
                                parent_id:
                                    type: string
                                    format: uuid
                                    nullable: true
            responses:
                "200":
                    description: Tag updated.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Tag"
                "400":
                    description: parent_id references a non-existent tag or would create an error cycle.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error400"
                "404":
                    description: Tag not found.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error404"
                "500":
                    description: Internal server error.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error500"

        delete:
            operationId: deleteTag
            summary: Delete a tag
            description: >
                Deletes the tag and all its note_tags assignments. Child
                tags are orphaned to the top level.
            tags: [Tags]
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: string
                      format: uuid
            responses:
                "204":
                    description: Tag deleted. Child tags orphaned, not removed.
                "404":
                    description: Tag not found.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error404"
                "500":
                    description: Internal server error.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error500"
    /notes/{id}/tags:
        post:
            operationId: assignTag
            summary: Assign a tag to a note
            tags: [Tags]
            parameters:
                - $ref: "#/components/parameters/NoteId"
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required: [tag_id]
                            properties:
                                tag_id:
                                    type: string
                                    format: uuid
            responses:
                "200":
                    description: Tag was already assigned to this note. No change made.
                "201":
                    description: Tag newly assigned to this note.
                "404":
                    description: Note or tag not found.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error404"
                "500":
                    description: Internal server error.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error500"
    /notes/{id}/tags/{tag_id}:
        delete:
            operationId: removeTag
            summary: Remove a tag assignment from a note
            tags: [Tags]
            parameters:
                - $ref: "#/components/parameters/NoteId"
                - name: tag_id
                  in: path
                  required: true
                  schema:
                      type: string
                      format: uuid
            responses:
                "204":
                    description: Assignment removed.
                "404":
                    description: Note, tag or the assignment itself does not exist.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error404"
                "500":
                    description: Internal server error.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error500"

    #  LINKS
...